LDEV-1707: Converted marks to floats, to allow marks like 7.5. Also I18N'd the output of the marks so that they are shown as 1,234.5 or 1 234,5 depending on the user's locale.

+++ lams_tool_sbmt/db/sql/create_lams_tool_sbmt.sql
@@ -50,7 +50,7 @@ create table tl_lasbmt11_instruction_files (
 create table tl_lasbmt11_report (
    report_id bigint not null auto_increment,
    comments varchar(250),
-   marks bigint,
+   marks float,
    date_marks_released datetime,
    primary key (report_id)
 )TYPE=InnoDB;
